+++ src/Data/Aeson/Types/FromJSON.hs
@@ -1542,6 +1542,10 @@ instance (FromJSON a, FromJSON b) => FromJSON (Either a b) where
 instance FromJSON Void where
     parseJSON _ = fail "Cannot parse Void"
 
+-- | @since 2.1.2.0
+instance FromJSONKey Void where
+    fromJSONKey = FromJSONKeyTextParser $ \_ -> fail "Cannot parse Void"
+
 instance FromJSON Bool where
     parseJSON (Bool b) = pure b
     parseJSON v = typeMismatch "Bool" v